+++ CHANGELOG.md
@@ -8,8 +8,23 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- None
+
+### Changed
+
+- None
+
+### Fixed
+
+- None
+
+## [6.2.0] - 2025-02-22
+
+### Added
+
+- Add support for SF Symbols 6.2 (By [Steven Sorial](https://github.com/StevenSorial))
 - Add support for SF Symbols 6 (By [Rod Brown](https://github.com/RodBrown1988))
-- Added initializers for SwiftUI Button.
+- Added more convenience initializers.
 
 ### Changed